fix transforming all tiles
Before this fix, if you already had objects present before importing the tiles, they'd be transformed as well as the imported tiles. This fix ensures that only the imported tiles are transformed. This is probably really inefficient if you're importing into a project with a large number of pre-existing objects, so it might be beneficial to look into ways to improve the performance.