masking data to/from saved files
We can now save masking data to files, and when we reopen them, the saved data is detected and the option activated in the GUI to import it. If the user chooses to do so, the tiles are first loaded as usual, and then the masking is performed automatically once all tiles have been loaded.