renaming masking objects; remove redundant option
You can now rename the masking objects through the window opened when running the "Mask" command. I also removed the now redundant option in the main window as to whether the user wants the masking data to be loaded in from file. It now automatically is, and they can manage this through the Mask command window if they don't want it.